feat: Add support for analog_input in miscellaneous panel
Klipper's analog_input type can have a value and a unit, and so is quite
similar to a MoonrakerSensor, but can not have multiple values per sensor.

This adds a MiscellaneousSensor panel type that is included in the
MiscellaneousPanel, and the getters necessary to connect analog_input
objects in to it.

Signed-off-by: Russell Cloran <rcloran@gmail.com>